# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API Endpoint for handling organizational financial reimbursement requests",
        "title": "Financial Request Handler API",
        "contact": {
            "name": "APP Support",
            "email": "app_support@norainc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/me/vehicle": {
            "put": {
                "description": "edits information about a specific vehicle for a logged in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "no-cache"
                ],
                "summary": "edit a vehicle",
                "operationId": "edit-vehicle",
                "parameters": [
                    {
                        "description": "edited vehicle information",
                        "name": "vehicle-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.VehicleRes"
                        }
                    }
                }
            },
            "post": {
                "description": "adds vehicle information to a logged in user's account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "no-cache"
                ],
                "summary": "add a vehicle",
                "operationId": "add-vehicle",
                "parameters": [
                    {
                        "description": "user's vehicle information",
                        "name": "vehicle-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VehicleInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.VehicleRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "removes vehicle information from a logged in user's account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "no-cache"
                ],
                "summary": "remove a vehicle",
                "operationId": "remove-vehicle",
                "parameters": [
                    {
                        "description": "deleted vehicle information",
                        "name": "vehicle-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.VehicleRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Vehicle": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.VehicleInput": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "responses.VehicleRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Vehicle"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        }
    }
}